requirements.txt
pbr>=0.6,!=0.7,<1.0
SQLAlchemy>=0.7.8,!=0.9.5,<=0.9.99
anyjson>=0.3.3
argparse
boto>=2.12.0,!=2.13.0
eventlet>=0.13.0
Jinja2
kombu>=2.4.8
lxml>=2.3
Routes>=1.12.3,!=2.0
WebOb>=1.2.3
greenlet>=0.3.2
PasteDeploy>=1.5.0
Paste
sqlalchemy-migrate>=0.9.1
netaddr>=0.7.6
suds>=0.4
paramiko>=1.13.0
posix_ipc
pyasn1
Babel>=1.3
iso8601>=0.1.9
jsonschema>=2.0.0,<3.0.0
python-cinderclient>=1.0.7
python-neutronclient>=2.3.5,<3
python-glanceclient>=0.13.1
python-keystoneclient>=0.9.0
six>=1.7.0
stevedore>=0.14
websockify>=0.5.1,<0.6
wsgiref>=0.1.2
oslo.config>=1.4.0.0a1  # FIXME(markmc): update when generator is released
oslo.rootwrap
oslotest
pycadf>=0.5.1
oslo.messaging>=1.3.0
